removed new hitching post, as it will cause existing ones to delete. made new hitching post for dungeons, etc.
